backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
selenium==4.15.2
webdriver-manager==4.0.1
pydantic>=2.5.2
pyotp
beautifulsoup4==4.12.2
cachetools>=5.3.2
requests>=2.28.0
pdfplumber==0.10.3
httpx>=0.27.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
supabase>=2.0.0
google-auth==2.27.0
dateparser>=1.2.0
